requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
APScheduler==3.6.3
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
build==1.3.0
cachetools==4.2.2
ccxt==4.5.22
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
coincurve==21.0.0
cryptography==45.0.6
curl_cffi==0.13.0
dash==3.2.0
dotenv==0.9.9
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.1
frozendict==2.4.6
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.4
multitasking==0.0.12
narwhals==2.1.1
nest-asyncio==1.6.0
numpy==1.24.4
packaging==25.0
pandas==2.1.4
peewee==3.18.2
platformdirs==4.5.0
plotly==6.3.0
propcache==0.3.2
protobuf==6.33.0
pycares==4.10.0
pycparser==2.22
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
retrying==1.4.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
TA-Lib==0.6.5
tomli==2.2.1
tornado==6.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.2
urllib3==1.26.18
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
yfinance==0.2.66
zipp==3.23.0
